refactor: water를 모듈로 단순화
